* [ios] move gradient type transform function to WXConvert.h
